add_subprotocol(): mimic argument matching for vectors in general *
This is an approach to consistently document all allowed values of
optional arguments
in the \usage{} section, the first value being the default, as is commonly done for
string arguments with the help of match.arg().
This code lets the programmer to pass allowed values in the function definition,
which can then be seen in the Usage section, just like for characters.